pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "motionmatch"
version = "0.1.0"
description = "Physics-based video search engine using V-JEPA 2"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "MotionMatch Team"}
]
keywords = ["video", "search", "motion", "vjepa", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "fastapi>=0.104.0",
    "celery>=5.3.0",
    "pymilvus>=2.3.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
preprocessing = [
    "scenedetect>=0.6.2",
    "ultralytics>=8.0.0",
]

[project.scripts]
motionmatch-api = "motionmatch.api.main:main"
motionmatch-worker = "motionmatch.workers.celery_app:main"

[project.urls]
Homepage = "https://github.com/iamvibhorsingh/motionmatch"
Repository = "https://github.com/iamvibhorsingh/motionmatch"
Issues = "https://github.com/iamvibhorsingh/motionmatch/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
